Paypal Integration

+++ account/views.py
@@ -1,11 +1,11 @@
 from django.shortcuts import redirect, render
 from .forms import CreateUserForm, LoginForm, UpdateUserForm
+from payment.models import OrderItem, ShippingAddress
+from payment.forms import ShippingForm
 from .token import user_tokenizer_generate
-
 # from payment.forms import ShippingForm
 # from payment.models import ShippingAddress
 # from payment.models import Order, OrderItem
-
 from django.contrib.auth.models import User
 from django.contrib.sites.shortcuts import get_current_site
 from django.template.loader import render_to_string
@@ -14,7 +14,6 @@ from django.utils.http import urlsafe_base64_decode, urlsafe_base64_encode
 from django.contrib.auth.models import auth
 from django.contrib.auth import authenticate
 from django.contrib.auth.decorators import login_required
-
 from django.contrib import messages
 
 
@@ -26,7 +25,6 @@ def register(request):
             user = form.save()
             user.is_active = False
             user.save()
-
             # Email verification setup (template)
             current_site = get_current_site(request)
             subject = 'Account verification email'
@@ -38,7 +36,6 @@ def register(request):
             })
             user.email_user(subject=subject, message=message)
             return redirect('email-verification-sent')
-
     context = {'form': form}
     return render(request, 'account/registration/register.html', context=context)
 
@@ -46,7 +43,6 @@ def register(request):
 def email_verification(request, uidb64, token):
     unique_id = force_str(urlsafe_base64_decode(uidb64))
     user = User.objects.get(pk=unique_id)
-
     if user and user_tokenizer_generate.check_token(user, token):
         user.is_active = True
         user.save()
@@ -74,11 +70,10 @@ def my_login(request):
         if form.is_valid():
             username = request.POST.get('username')
             password = request.POST.get('password')
-
             user = authenticate(request, username=username, password=password)
             if user is not None:
                 auth.login(request, user)
-                return redirect("dashboard")
+                return redirect("store")
     context = {'form': form}
     return render(request, 'account/my-login.html', context=context)
 
@@ -92,7 +87,6 @@ def user_logout(request):
                 del request.session[key]
     except KeyError:
         pass
-
     messages.success(request, "Logout success")
     return redirect("store")
 
@@ -110,7 +104,6 @@ def profile_management(request):
         user_form = UpdateUserForm(request.POST, instance=request.user)
         if user_form.is_valid():
             user_form.save()
-
             messages.info(request, "Update success!")
             return redirect('dashboard')
     context = {'user_form': user_form}
@@ -120,69 +113,43 @@ def profile_management(request):
 @login_required(login_url='my-login')
 def delete_account(request):
     user = User.objects.get(id=request.user.id)
-    if request.method == 'POST':
+    if user and request.method == 'POST':
         user.delete()
-
         messages.error(request, "Account deleted")
         return redirect('store')
-
     return render(request, 'account/delete-account.html')
 
-
-# # Shipping view
-# @login_required(login_url='my-login')
-# def manage_shipping(request):
-
-#     try:
-
-#         # Account user with shipment information
-
-#         shipping = ShippingAddress.objects.get(user=request.user.id)
-
-#     except ShippingAddress.DoesNotExist:
-
-#         # Account user with no shipment information
-
-#         shipping = None
-
-#     form = ShippingForm(instance=shipping)
-
-#     if request.method == 'POST':
-
-#         form = ShippingForm(request.POST, instance=shipping)
-
-#         if form.is_valid():
-
-#             # Assign the user FK on the object
-
-#             shipping_user = form.save(commit=False)
-
-#             # Adding the FK itself
-
-#             shipping_user.user = request.user
-
-#             shipping_user.save()
-
-#             messages.info(request, "Update success!")
-
-#             return redirect('dashboard')
-
-#     context = {'form': form}
-
-#     return render(request, 'account/manage-shipping.html', context=context)
+# Shipping view
 
 
-# @login_required(login_url='my-login')
-# def track_orders(request):
+@login_required(login_url='my-login')
+def manage_shipping(request):
+    try:
+        # Account user with shipment information
+        shipping = ShippingAddress.objects.get(user=request.user.id)
+    except ShippingAddress.DoesNotExist:
+        # Account user with no shipment information
+        shipping = None
+    form = ShippingForm(instance=shipping)
+    if request.method == 'POST':
+        form = ShippingForm(request.POST, instance=shipping)
+        if form.is_valid():
+            # Assign the user FK on the object
+            shipping_user = form.save(commit=False)
+            # Adding the FK itself
+            shipping_user.user = request.user
+            shipping_user.save()
+            messages.info(request, "Update success!")
+            return redirect('dashboard')
+    context = {'form': form}
+    return render(request, 'account/manage-shipping.html', context=context)
 
-#     try:
 
-#         orders = OrderItem.objects.filter(user=request.user)
-
-#         context = {'orders': orders}
-
-#         return render(request, 'account/track-orders.html', context=context)
-
-#     except:
-
-#         return render(request, 'account/track-orders.html')
+@login_required(login_url='my-login')
+def track_orders(request):
+    try:
+        orders = OrderItem.objects.filter(user=request.user)
+        context = {'orders': orders}
+        return render(request, 'account/track-orders.html', context=context)
+    except:
+        return render(request, 'account/track-orders.html')
